create helper function file for all no jspsych related functions
