+++ src/EmotionRecognitionTask.ts
@@ -4,6 +4,7 @@ import HtmlKeyboardResponsePlugin from "@jspsych/plugin-html-keyboard-response";
 import HtmlButtonResponse from '@jspsych/plugin-html-button-response';
 import PreloadPlugin from "@jspsych/plugin-preload";
 import { JsPsych } from "/runtime/v1/jspsych@8.x";
+import { addBootstrapScripts, addContinueButton, createContinueButtonDiv, revealEmotionButtons } from "./helperFunctions";
 
 var jsPsych = initJsPsych();
 
@@ -419,62 +420,9 @@ export default function emotionRecognitionTask() {
   timeline.push(videoCheckWithButtons);
   jsPsych.run(timeline);
 
-  const addBootstrapScripts = () => {
-    let link = document.createElement('link')
-    let ajaxScript = document.createElement('script')
-    let bootstrapScript = document.createElement('script')
-
-    link.rel = "stylesheet"
-    link.href = "https://maxcdn.bootstrapcdn.com/bootstrap/3.4.1/css/bootstrap.min.css"
-
-    ajaxScript.src = "https://ajax.googleapis.com/ajax/libs/jquery/3.7.1/jquery.min.js"
-    bootstrapScript.src = "https://maxcdn.bootstrapcdn.com/bootstrap/3.4.1/js/bootstrap.min.js"
-
-    document.head.appendChild(link)
-    document.head.appendChild(ajaxScript)
-    document.head.appendChild(bootstrapScript)
-  }
-
   function simulateKeyPress(jsPsych: JsPsych, key: string) {
     jsPsych.pluginAPI.keyDown(key);
     jsPsych.pluginAPI.keyUp(key);
-  }
-
-  const addContinueButton = () => {
-    let divContainer = document.createElement('div')
-    let continueButton = document.createElement('button')
-
-    divContainer.style.alignContent = "center"
-
-    continueButton.style.alignContent = "center"
-    continueButton.style.justifyContent = "center"
-    continueButton.style.position = "flex"
-    continueButton.style.display = "None"
-    continueButton.style.margin = "0"
-    continueButton.className = "btn btn-primary"
-    continueButton.textContent = "Continue"
-    
-    return continueButton
-  }
-
-  const createContinueButtonDiv = (continueButton: HTMLButtonElement) => {
-    let continueButtonDiv = document.createElement('div')
-    continueButtonDiv.style.justifyContent = "center"
-    continueButtonDiv.style.alignItems = "center"
-    continueButtonDiv.style.display = "flex"
-    continueButtonDiv.appendChild(continueButton)
-
-    return continueButtonDiv
-
-  }
-
-  const revealEmotionButtons = () => {
-
-    let emotionButtons = document.getElementsByName('custom-button-div')
-
-    for(let i = 0 ; i < emotionButtons.length; i ++){
-      emotionButtons[i].style.display = 'flex'
-    }
   }
 
 }
